+++ patches/src/main/kotlin/app/jptt/patches/shared/Constants.kt
@@ -11,6 +11,7 @@ internal object Constants {
     const val EXTENSION_PRELOAD_CLASS = "Lapp/jptt/extension/PreloadArticleImagesPatch;"
     const val EXTENSION_FILE_PROVIDER_CLASS = "Lapp/jptt/extension/FileProviderAuthorityPatch;"
     const val EXTENSION_TERMINAL_ESCAPE_CLASS = "Lapp/jptt/extension/TerminalEscapePatch;"
+    const val EXTENSION_SEARCH_HISTORY_CLASS = "Lapp/jptt/extension/SearchHistoryLayoutPatch;"
 
     const val ARTICLE_FRAGMENT_CLASS = "Lcom/joshua/jptt/ArticleFragment;"
     const val BOARD_FRAGMENT_CLASS = "Lcom/joshua/jptt/BoardFragment;"